서류함 조회 완료

+++ src/main/java/com/wisehr/wisehr/alarmAndMessage/dto/SendMessengerDTO.java
@@ -13,6 +13,6 @@ public class SendMessengerDTO {
     private String msgCode;
     private Date msgDate;
     private String msgContents;
-    private int memCode;
+    private Long memCode;
     private String msgDeleteStatus;
 }
